put fit and model in one file to avoid import issues
